refactor(stella-tui): the turn in flight moves to model/turn.rs, off model.rs's ceiling
The opening rule's `TurnOpening` pushed `model.rs` to 1532 lines, 32 over
the guard. Split rather than baselined: AGENTS.md's remedy for a file
crossing the limit is a sibling submodule, and the baseline covers only
files that predate the guard.

`Hud` moves with it, because the two are the ends of one thing — `Hud` is
the turn's live state, `TurnOpening` is the stamp that state leaves on the
transcript when a turn opens — and moving only the 44 new lines would
have landed at 1490, which is a ceiling to hit again next month rather
than headroom. The seam is the one `file_state` and `recall` already cut:
a `TranscriptEntry`'s supporting types live beside `model.rs`, not in it.

Both are re-exported from `crate::model`, so no call site changed. 1430
lines now.

Refs #4124
